move GetAllGroups and performer from the metrics code to the general model code so it can be used anywhere
